got rid of interaction layer in favor of more complete dismod objects
