+++ src/cascade/model/session.py
@@ -4,9 +4,7 @@ import pandas as pd
 
 from cascade.core import getLoggers
 from cascade.model import Model
-from cascade.model.interaction import (
-    prepare_init, prepare_fit, run_dismod, prepare_predict, amend_data_input
-)
+from cascade.model.data_read_write import amend_data_input, _point_age_time_to_interval
 from cascade.model.object_wrapper import ObjectWrapper
 
 CODELOG, MATHLOG = getLoggers(__name__)
@@ -117,7 +115,7 @@ class Session:
                 raise ValueError(f"Model and initial guess are misaligned: {misalignment}.")
         self.setup_model_for_fit(model, data, initial_guess)
         # else use the one generated by the call to init, coming from the mean.
-        dm_out, dm_err = run_dismod(self._objects, ["fit", fit_level])
+        dm_out, dm_err = self._objects.run_dismod(["fit", fit_level])
         return FitResult(self._objects, self._objects.fit_var, dm_out, dm_err)
 
     def setup_model_for_fit(self, model, data=None, initial_guess=None):
@@ -130,17 +128,22 @@ class Session:
             data (pd.DataFrame|None): Can be None.
             initial_guess (Var|None): Initial values, can be None.
         """
-        prepare_init(
-            self._objects, self._locations, self._parent_location, model, data)
+        self._objects.locations = self._locations
+        self._objects.parent_location_id = self._parent_location
+        self._objects.model = model
+        # The data has to be in there before init in order to build
+        # the data subset table.
+        self._objects.data = data
         self._objects.set_option(**self._options)
         if self._age_extents:
             self._objects.age_extents = self._age_extents
-        run_dismod(self._objects, "init")
+        self._objects.run_dismod("init")
         if model.scale_set_by_user:
-            scale = model.scale
-        else:
-            scale = None
-        prepare_fit(self._objects, initial_guess, scale)
+            self._objects.scale_var = model.scale
+        # else Dismod uses the initial guess as the scale.
+        if initial_guess is not None:
+            self._objects.start_var = initial_guess
+        # else Dismod-AT uses the distribution means as the start_var.
 
     def predict(self, var, avgint, parent_location, weights=None, covariates=None):
         """Given rates, calculated the requested average integrands.
@@ -171,8 +174,10 @@ class Session:
         self._check_vars(var)
         model = Model.from_var(var, parent_location, weights=weights, covariates=covariates)
         self.setup_model_for_fit(model)
-        prepare_predict(self._objects, avgint, var)
-        run_dismod(self._objects, ["predict", "truth_var"])
+        avgint = _point_age_time_to_interval(avgint)
+        self._objects.avgint = avgint
+        self._objects.truth_var = var
+        self._objects.run_dismod(["predict", "truth_var"])
         predicted, not_predicted = self._objects.predict
         return predicted, not_predicted
 
@@ -203,7 +208,7 @@ class Session:
                 raise ValueError(f"Model and fit var are misaligned: {misalignment}.")
         self.setup_model_for_fit(model, data, fit_var)
         self._objects.truth_var = fit_var
-        run_dismod(self._objects, ["simulate", simulate_count])
+        self._objects.run_dismod(["simulate", simulate_count])
         amended_data = amend_data_input(data)
         return SimulateResult(self._objects, simulate_count, model, amended_data)
 
@@ -216,7 +221,7 @@ class Session:
         Returns:
             DismodGroups[Var] with multiple samples.
         """
-        run_dismod(self._objects, ["sample", "simulate", simulate_result.count])
+        self._objects.run_dismod(["sample", "simulate", simulate_result.count])
         return self._objects.samples
 
     def set_option(self, **kwargs):
